requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==3.0.3
numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.4.0
ultralytics==8.2.51
